+++ packages/cojson/src/localNode.ts
@@ -26,9 +26,9 @@ import {
     Account,
     AccountMeta,
     accountHeaderForInitialAgentSecret,
-    GeneralizedControlledAccount,
+    ControlledAccountOrAgent,
     ControlledAccount,
-    AnonymousControlledAccount,
+    ControlledAgent,
     AccountID,
     Profile,
     AccountMigration,
@@ -52,7 +52,7 @@ export class LocalNode {
     /** @internal */
     coValues: { [key: RawCoID]: CoValueState } = {};
     /** @category 3. Low-level */
-    account: GeneralizedControlledAccount;
+    account: ControlledAccountOrAgent;
     /** @category 3. Low-level */
     currentSessionID: SessionID;
     /** @category 3. Low-level */
@@ -60,7 +60,7 @@ export class LocalNode {
 
     /** @category 3. Low-level */
     constructor(
-        account: GeneralizedControlledAccount,
+        account: ControlledAccountOrAgent,
         currentSessionID: SessionID
     ) {
         this.account = account;
@@ -68,7 +68,7 @@ export class LocalNode {
     }
 
     /** @category 2. Node Creation */
-    static withNewlyCreatedAccount<
+    static async withNewlyCreatedAccount<
         P extends Profile = Profile,
         R extends CoMap = CoMap,
         Meta extends AccountMeta = AccountMeta
@@ -80,15 +80,15 @@ export class LocalNode {
         name: string;
         migration?: AccountMigration<P, R, Meta>;
         initialAgentSecret?: AgentSecret;
-    }): {
+    }): Promise<{
         node: LocalNode;
         accountID: AccountID;
         accountSecret: AgentSecret;
         sessionID: SessionID;
-    } {
+    }> {
         const throwawayAgent = newRandomAgentSecret();
         const setupNode = new LocalNode(
-            new AnonymousControlledAccount(throwawayAgent),
+            new ControlledAgent(throwawayAgent),
             newRandomSessionID(getAgentID(throwawayAgent))
         );
 
@@ -108,7 +108,7 @@ export class LocalNode {
         );
 
         if (migration) {
-            migration(accountOnNodeWithAccount, profile as P);
+            await migration(accountOnNodeWithAccount, profile as P, nodeWithAccount);
         }
 
         nodeWithAccount.account = new ControlledAccount(
@@ -156,7 +156,7 @@ export class LocalNode {
         migration?: AccountMigration<P, R, Meta>;
     }): Promise<LocalNode> {
         const loadingNode = new LocalNode(
-            new AnonymousControlledAccount(accountSecret),
+            new ControlledAgent(accountSecret),
             newRandomSessionID(accountID)
         );
 
@@ -194,9 +194,10 @@ export class LocalNode {
         const profile = await node.load(profileID);
 
         if (migration) {
-            migration(
+            await migration(
                 controlledAccount as ControlledAccount<P, R, Meta>,
-                profile as P
+                profile as P,
+                node
             );
             node.account = new ControlledAccount(
                 controlledAccount.core,
@@ -370,14 +371,14 @@ export class LocalNode {
         const groupAsInvite = expectGroup(
             group.core
                 .testWithDifferentAccount(
-                    new AnonymousControlledAccount(inviteAgentSecret),
+                    new ControlledAgent(inviteAgentSecret),
                     newRandomSessionID(inviteAgentID)
                 )
                 .getCurrentContent()
         );
 
         groupAsInvite.addMemberInternal(
-            this.account.id,
+            this.account,
             inviteRole === "adminInvite"
                 ? "admin"
                 : inviteRole === "writerInvite"
@@ -439,7 +440,7 @@ export class LocalNode {
         let account = expectGroup(
             this.createCoValue(accountHeaderForInitialAgentSecret(agentSecret))
                 .testWithDifferentAccount(
-                    new AnonymousControlledAccount(agentSecret),
+                    new ControlledAgent(agentSecret),
                     newRandomSessionID(accountAgentID)
                 )
                 .getCurrentContent()
@@ -518,7 +519,7 @@ export class LocalNode {
             );
         }
 
-        return new Account(coValue).getCurrentAgentID();
+        return new Account(coValue).currentAgentID();
     }
 
     async resolveAccountAgentAsync(
@@ -553,7 +554,7 @@ export class LocalNode {
             );
         }
 
-        return new Account(coValue).getCurrentAgentID();
+        return new Account(coValue).currentAgentID();
     }
 
     /**
@@ -596,7 +597,7 @@ export class LocalNode {
 
     /** @internal */
     testWithDifferentAccount(
-        account: GeneralizedControlledAccount,
+        account: ControlledAccountOrAgent,
         currentSessionID: SessionID
     ): LocalNode {
         const newNode = new LocalNode(account, currentSessionID);