+++ examples/twit-stresstest/CHANGELOG.md
@@ -1,5 +1,24 @@
 # twit-stresstest
 
+## 0.2.0
+
+### Minor Changes
+
+- Make addMember and removeMember take loaded Accounts instead of just IDs
+
+### Patch Changes
+
+- Updated dependencies
+  - jazz-nodejs@0.6.0
+
+## 0.1.1
+
+### Patch Changes
+
+- Allow account migrations to be async
+- Updated dependencies
+  - jazz-nodejs@0.5.3
+
 ## 0.1.0
 
 ### Minor Changes