+++ packages/cojson-simple-sync/CHANGELOG.md
@@ -1,5 +1,17 @@
 # cojson-simple-sync
 
+## 0.6.0
+
+### Minor Changes
+
+- Make addMember and removeMember take loaded Accounts instead of just IDs
+
+### Patch Changes
+
+- Updated dependencies
+  - cojson@0.6.0
+  - cojson-storage-sqlite@0.5.2
+
 ## 0.5.0
 
 ### Minor Changes