+++ packages/cojson-storage-sqlite/CHANGELOG.md
@@ -1,5 +1,12 @@
 # cojson-storage-sqlite
 
+## 0.5.2
+
+### Patch Changes
+
+- Updated dependencies
+  - cojson@0.6.0
+
 ## 0.5.1
 
 ### Patch Changes