+++ packages/cojson-transport-nodejs-ws/CHANGELOG.md
@@ -1,5 +1,12 @@
 # cojson-transport-nodejs-ws
 
+## 0.5.1
+
+### Patch Changes
+
+- Updated dependencies
+  - cojson@0.6.0
+
 ## 0.5.0
 
 ### Minor Changes