+++ packages/cojson/CHANGELOG.md
@@ -1,5 +1,23 @@
 # cojson
 
+## 0.6.1
+
+### Patch Changes
+
+- Provide localNode to AccountMigrations
+
+## 0.6.0
+
+### Minor Changes
+
+- Make addMember and removeMember take loaded Accounts instead of just IDs
+
+## 0.5.2
+
+### Patch Changes
+
+- Allow account migrations to be async
+
 ## 0.5.1
 
 ### Patch Changes