+++ packages/jazz-browser/CHANGELOG.md
@@ -1,5 +1,26 @@
 # jazz-browser
 
+## 0.6.0
+
+### Minor Changes
+
+- Make addMember and removeMember take loaded Accounts instead of just IDs
+
+### Patch Changes
+
+- Updated dependencies
+  - cojson-storage-indexeddb@0.6.0
+  - jazz-autosub@0.6.0
+  - cojson@0.6.0
+
+## 0.5.1
+
+### Patch Changes
+
+- Allow account migrations to be async
+- Updated dependencies
+  - cojson@0.5.2
+
 ## 0.5.0
 
 ### Minor Changes