+++ packages/jazz-react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # jazz-react
 
+## 0.5.1
+
+### Patch Changes
+
+- Updated dependencies
+  - jazz-browser@0.6.0
+  - cojson@0.6.0
+
 ## 0.5.0
 
 ### Minor Changes